Move prerender function out into its own file
