+++ webpack.server.config.js
@@ -1,5 +1,5 @@
-var webpack = require("webpack");
-var path = require("path");
+var webpack = require('webpack');
+var path = require('path');
 var fs = require('fs');
 
 // load env vars first
